fix type issueNumber field from Long to Int

+++ selfouettie-server/src/main/scala/net/yoshinorin/selfouettie/models/db/IssueCommentEvents.scala
@@ -6,7 +6,7 @@ case class IssueCommentEvents(
   eventId: Long,
   userName: String,
   repositoryId: Long,
-  issueNumber: Long,
+  issueNumber: Int,
   action: String,
   createdAt: Long
 ) extends BaseEvent[IssueCommentEvents]
